Game info screen: Show list of any plugins that will be loaded on start of the game
