+++ Core/HLE/Plugins.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include <cstdint>
+#include <string>
 #include "Input/KeyCodes.h"
 
 class PointerWrap;
@@ -33,6 +34,21 @@ void Unload();
 void DoState(PointerWrap &p);
 
 bool HasEnabled();
+
+enum class PluginType {
+	INVALID = 0,
+	PRX,
+};
+
+struct PluginInfo {
+	PluginType type;
+	std::string name;
+	std::string filename;  // PSP-space path. So we can't use a Path object.
+	int version;
+	uint32_t memory;
+};
+
+std::vector<PluginInfo> FindPlugins(const std::string &gameID, const std::string &lang);
 
 void SetKey(int key, uint8_t value);
 uint8_t GetKey(int key);
