+++ assets/lang/pt_BR.ini
@@ -1274,6 +1274,7 @@ Change CPU Clock = Mudar o clock da CPU do PSP emulado (instável)
 CPU Core = Núcleo da CPU
 Dynarec/JIT (recommended) = Dynarec/JIT (recomendado)
 JIT using IR = JIT using IR
+Plugins = Plugins
 Loaded plugin: %1 = Plugin carregado: %1
 Memory Stick folder = Pasta do cartão de memória
 Memory Stick size = Tamanho do cartão de memória
